requirements.txt
pip~=24.2
wheel~=0.41.2
setuptools~=68.2.0
filelock~=3.16.0
requests~=2.32.3
PyYAML~=6.0.2
numpy~=1.24.4
charset-normalizer~=3.3.2
regex~=2024.9.11
fsspec~=2024.9.0
certifi~=2024.8.30
urllib3~=2.2.2
idna~=3.8
packaging~=24.1
tokenizers~=0.19.1
safetensors~=0.4.5
tqdm~=4.66.5
transformers~=4.44.2
huggingface-hub~=0.24.7
hpack~=4.0.0
hyperframe~=6.0.1
h2~=4.1.0
GitPython~=3.1.43
gitdb~=4.0.11
h11~=0.12.0
future~=1.0.0
pillow~=10.4.0
attrs~=24.2.0
protobuf~=3.19.6
rsa~=4.9
pyasn1~=0.6.1
absl-py~=2.1.0
dill~=0.3.8
gast~=0.4.0
grpcio~=1.66.1
h5py~=3.11.0
Mako~=1.3.5
MarkupSafe~=2.1.5
pytz~=2024.2
keras~=2.9.0
multidict~=6.1.0
zipp~=3.20.1
anyio~=3.7.1
sniffio~=1.3.1
libclang~=18.1.1
werkzeug~=3.0.4
click~=8.1.7
jinja2~=3.1.4
blinker~=1.8.2
itsdangerous~=2.2.0
smmap~=5.0.1
rfc3986~=1.5.0
httpcore~=0.13.7
tensorflow~=2.9.3
six~=1.16.0
scipy~=1.10.1
pandas~=1.5.3
wrapt~=1.16.0
google-pasta~=0.2.0
matplotlib~=3.7.5
typer~=0.3.2
docker~=7.1.0
ffmpeg-python~=0.2.0
aiohttp~=3.10.5
yarl~=1.11.1
pyasn1-modules~=0.4.1
cachetools~=5.5.0
joblib~=1.4.2
cloudpickle~=3.0.0
mlflow~=2.16.0
mlflow-skinny~=2.16.0
datasets~=3.0.0
SQLAlchemy~=2.0.34
alembic~=1.13.2
pyarrow~=17.0.0
importlib-metadata~=8.4.0
sqlparse~=0.5.1
scikit-learn~=1.3.2
flask~=3.0.3
graphene~=3.3
graphql-core~=3.2.4
databricks-sdk~=0.32.1
python-dateutil~=2.9.0.post0
xxhash~=3.5.0
async-timeout~=4.0.3
gunicorn~=23.0.0
aiosignal~=1.3.1
frozenlist~=1.4.1
aiohappyeyeballs~=2.4.0
importlib-resources~=6.4.5
tzdata~=2024.1
threadpoolctl~=3.5.0
multiprocess~=0.70.16
evaluate~=0.4.3
graphql-relay~=3.2.0
aniso8601~=9.0.1
Deprecated~=1.2.14
greenlet~=3.1.0
Markdown~=3.7
oauthlib~=3.2.2
spleeter~=2.4.0
httpx~=0.19.0
contourpy~=1.1.1
fonttools~=4.53.1
pyparsing~=3.1.4
termcolor~=2.4.0
cycler~=0.12.1
kiwisolver~=1.4.7
opt-einsum~=3.3.0
tensorboard~=2.9.1
flatbuffers~=1.12
astunparse~=1.6.3
google-auth-oauthlib~=0.4.6
Keras-Preprocessing~=1.1.2
requests-oauthlib~=2.0.0
pydub~=0.25.1